ferme-backend/requirements.txt
Django>=5.0
gunicorn
psycopg2-binary
dj-database-url
djangorestframework
django-cors-headers
django-jazzmin
Pillow            
python-telegram-bot
whitenoise
django-storages